some basic movement

+++ src/intf/player_manager.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 extern int prev_x, prev_y;
-extern float pos_x, pos_y;
+extern int pos_x, pos_y;
 
 void move_hori(char _input);
 void move_vert(char _input);
